add pick function
