+++ client/.env
@@ -1,1 +1,1 @@
-SERVER_ADDRESS=
+SERVER_ADDRESS=http://ec2-3-35-218-199.ap-northeast-2.compute.amazonaws.com
