+++ client/src/components/Picklist.js
@@ -4,15 +4,37 @@ import styled from "styled-components";
 
 const Wrapper = styled.div`
   border: 3px solid yellowgreen;
+  width: 100%;
+  height: 100%;
+  flex-wrap: wrap;
+  overflow-y: auto;
+  -ms-overflow-style: none; /* for Internet Explorer, Edge */
+  scrollbar-width: none; /* for Firefox */
+  &::-webkit-scrollbar {
+    display: none; /* for Chrome, Safari, and Opera */
+  }
 `;
-const Picklist = () => {
-  return (
-    <Wrapper>
-      <Pick />
-      <Pick />
-      <Pick />
-    </Wrapper>
-  );
+const Picklist = ({ festivalData, pickItems, removePick }) => {
+  if (festivalData === null) {
+    return <div>데이터를 받아오는 중</div>;
+  } else {
+    const renderedItems = festivalData.filter(
+      (ele) => pickItems.map((el) => el.festival_Id).indexOf(ele.id) > -1
+    );
+    return (
+      <div>
+        {pickItems.length === 0 ? (
+          <div>찜한 축제가 없습니다</div>
+        ) : (
+          <Wrapper>
+            {renderedItems.map((item, idx) => {
+              return <Pick key={idx} item={item} removePick={removePick} />;
+            })}
+          </Wrapper>
+        )}
+      </div>
+    );
+  }
 };
 
 export default Picklist;
