+++ client/src/pages/Pickpage.js
@@ -4,15 +4,23 @@ import styled from "styled-components";
 
 const Wrapper = styled.div`
   display: flex;
-  justify-content: space-between;
-  width: 60vw;
+  flex-direction: column;
+  width: 80vw;
+  height: 80vh;
+  border: 1px solid black;
 `;
 
-const Pickpage = () => {
+const Pickpage = ({ festivalData, pickItems, removePick }) => {
   return (
     <Wrapper>
-      <div>My Pick!</div>
-      <Picklist />
+      <div>
+        <h1>My Pick!</h1>
+      </div>
+      <Picklist
+        removePick={removePick}
+        festivalData={festivalData}
+        pickItems={pickItems}
+      />
     </Wrapper>
   );
 };
